refactor: Extract interface for SouvenirRepository
